fix: move vault to top conduit
vault is now configured for the whole app rather than at the level of
individual integrations.  Created encrypted binary type that can also be
reused and removed the custom type in Sage and references to this type
in the Sage Endpoint schema.

Changed configuration of vault to rely on onepassword command line tool.
The vault is now configured when it's genserver starts.  Added
descriptive errors to help when the vault fails to start.