fix: requesting access token now handles the case where we receive a new refresh token.
The QB API will periodically, but not predictably, return a new refresh
token in addition to an access token.  When it does the previously
issued refresh token is expired.  Added logic to endpoints and access
token module to handle this.

Also created an InvalidApiResponseError exception struct to represent
cases where an API endpoint returns data we can not parse or process.